backend/requirements.txt
anyio==3.7.0
asn1crypto==1.5.1
asttokens==2.2.1
bcrypt==4.0.1
bokeh==3.1.1
boto3==1.26.153
botocore==1.29.153
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
dask==2023.9.1
debugpy==1.6.7
decorator==5.1.1
distributed==2023.9.1
dnspython==2.3.0
python-dotenv==1.0.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
fastapi-cli==0.0.10
filelock==3.12.0
fsspec==2023.9.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.6
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.1.1
passlib[bcrypt]==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.1
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
pyarrow==10.0.1
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-docx==0.8.11
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pywin32==305
PyYAML==6.0
requests==2.31.0
rich==13.3.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.1
watchfiles==0.19.0
websockets==10.4
zipp==3.17.0
